requirements.txt
# === CORE BACKEND ===
fastapi==0.115.0
uvicorn==0.31.1


# === DOCUMENT PARSING ===
python-docx==1.1.0
pdfplumber==0.11.0
PyMuPDF==1.24.9
reportlab

# === AI / NLP / ML ===
transformers==4.48.2
torch==2.3.0
tensorflow==2.18.0
numpy==1.26.4
scikit-learn==1.5.2

# === UTILITIES ===
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.5
nltk==3.9.1
pydantic
python-dotenv
pymongo

# === OPTIONAL (but useful) ===
huggingface-hub==0.24.6
hf_xet==0.1.4
google-api-python-client==2.142.0

# === DEVELOPMENT ===
pandas==2.2.3
matplotlib==3.9.2
